Convert the plugin to a BitmapEffect
This is a prerequisite for allowing the plugin to write to areas
outside of the selection.
It also enables various other features for the 8bf filters, such as
anti-aliased/feathered selection masks and reading the document DPI.

+++ src/MaskSurface/MaskSurface.cs
@@ -32,7 +32,7 @@ namespace PSFilterLoad.PSApi
     {
         private int width;
         private int height;
-        private long stride;
+        private int stride;
         private MemoryBlock scan0;
         private bool disposed;
 
@@ -68,7 +68,7 @@ namespace PSFilterLoad.PSApi
         /// <summary>
         /// Gets the stride of the Surface.
         /// </summary>
-        public long Stride => stride;
+        public int Stride => stride;
 
         public MemoryBlock Scan0 => scan0;
